debrief-qa/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# HTTP client for ServiceTitan API
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# Templates
jinja2==3.1.3

# Date/time handling
python-dateutil==2.8.2

# For webhook signature verification
cryptography==42.0.0

# Authentication (Google OAuth)
authlib==1.3.0
itsdangerous==2.1.2
starlette[full]==0.35.1

# AI/ML for invoice review
google-generativeai>=0.8.0
